Fix integer casts in job runs table

+++ frontend/lib/database/job-runs-db.ts
@@ -123,9 +123,9 @@ export class JobRunsDB {
 
     const result = await db.query<{ id: number }>(
       `WITH row_data AS (
-         SELECT $1 AS job_id, $2 AS job_type, $3::integer AS company_id,
-                $4::integer AS output_file_id, $5 AS output_file_type,
-                $6::integer AS timeout, $7 AS source
+         SELECT $1 AS job_id, $2 AS job_type, $3 AS company_id,
+                $4 AS output_file_id, $5 AS output_file_type,
+                $6 AS timeout, $7 AS source
        )
        INSERT INTO job_runs (job_id, job_type, company_id, output_file_id, output_file_type, timeout, source, status)
        SELECT job_id, job_type, company_id, output_file_id, output_file_type, timeout, source, 'RUNNING'
@@ -210,7 +210,7 @@ export class JobRunsDB {
 
       const inserted = await tx.query<{ id: number }>(
         `WITH row_data AS (
-           SELECT $1 AS job_id, $2 AS job_type, $3::integer AS company_id, $4::integer AS timeout, $5 AS source
+           SELECT $1 AS job_id, $2 AS job_type, $3 AS company_id, $4 AS timeout, $5 AS source
          )
          INSERT INTO job_runs (job_id, job_type, company_id, timeout, source, status)
          SELECT job_id, job_type, company_id, timeout, source, 'RUNNING'
